Add flatpickr
Had to modify webpack config to get the css to work. I
just added another loader for css/sass files within
node_modules. I still have to explicitly include the
flatpickr.min.css in the bundle, but I am not sure if
this is normal configuration or not. Works for me.